refactor(operator): extract customer-yaml audit emission to its own module
The ADR 0049 escalation_model editor wiring pushed customer-yaml-editor.ts
to 502 counted lines, over the 500 ceiling (eslint max-lines). Move the
self-contained audit trio (buildAuditMetadata, recordCustomerYamlUpdateAudit,
CustomerYamlAuditMetadata) to customer-yaml-audit.ts — one-way dependency on
the editor's pure helpers, single consumer import updated.
