+++ src/lib/portal/operator/customer-yaml-editor.ts
@@ -682,47 +682,3 @@ function reconstructFromProjection(row: CustomerConfigRow): unknown {
     },
   }
 }
-
-// ============================================================================
-// Audit metadata
-// ============================================================================
-
-export interface CustomerYamlAuditMetadata {
-  changed_fields: string[]
-  before_hash: string
-  after_hash: string
-  actor_id: string
-}
-
-export function buildAuditMetadata(
-  before: EditableCustomerConfig,
-  after: EditableCustomerConfig,
-  actorId: string
-): CustomerYamlAuditMetadata {
-  return {
-    changed_fields: computeChangedFields(before, after),
-    before_hash: hashEditableConfig(before),
-    after_hash: hashEditableConfig(after),
-    actor_id: actorId,
-  }
-}
-
-/**
- * Emit the audit event to Worker tail logs. Mirrors the
- * `recordSendApprovedAudit` pattern in `send-approved.ts` (PR #960): a single
- * `console.info` line prefixed with `audit:customer_yaml_updated` so a
- * Hermes-side drain consumes it and persists to the per-customer D1.
- *
- * The audit fires even on validation failure — the attempt is itself
- * a recorded compliance event. Callers decide whether to emit on
- * failure (the route does, with `status: 'rejected'`).
- */
-export async function recordCustomerYamlUpdateAudit(payload: {
-  status: 'applied' | 'rejected'
-  customer_id: string
-  metadata: CustomerYamlAuditMetadata
-}): Promise<void> {
-  const line = JSON.stringify({ type: 'audit:customer_yaml_updated', ...payload })
-  console.info(line)
-  return Promise.resolve()
-}
